rag-playground-backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
langchain>=0.0.352
langchain-community>=0.0.10
langchain-groq>=0.0.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tiktoken>=0.5.2
numpy>=1.26.2
pydantic>=2.5.2
